+++ src/Phan/AST/FallbackUnionTypeVisitor.php
@@ -735,6 +735,25 @@ class FallbackUnionTypeVisitor extends KindVisitorImplementation
     }
 
     /**
+     * Visit a node with kind `\ast\AST_NULLSAFE_METHOD_CALL`.
+     *
+     * Conservatively try to infer the returned union type of calls such
+     * as $this?->someMethod(...)
+     *
+     * @param Node $node
+     * A node of the type indicated by the method name that we'd
+     * like to figure out the type that it produces.
+     *
+     * @return UnionType
+     * The set of types that are possibly produced by the
+     * given node
+     */
+    public function visitNullsafeMethodCall(Node $node): UnionType
+    {
+        return $this->visitMethodCall($node)->nullableClone();
+    }
+
+    /**
      * Visit a node with kind `\ast\AST_ASSIGN`
      *
      * @param Node $node