+++ src/Phan/AST/Visitor/KindVisitor.php
@@ -219,6 +219,11 @@ interface KindVisitor
     public function visitMethodCall(Node $node);
 
     /**
+     * Visit a node with kind `\ast\AST_NULLSAFE_METHOD_CALL`
+     */
+    public function visitNullsafeMethodCall(Node $node);
+
+    /**
      * Visit a node with kind `\ast\AST_NAME`
      */
     public function visitName(Node $node);
@@ -262,6 +267,11 @@ interface KindVisitor
      * Visit a node with kind `\ast\AST_PROP`
      */
     public function visitProp(Node $node);
+
+    /**
+     * Visit a node with kind `\ast\AST_NULLSAFE_PROP`
+     */
+    public function visitNullsafeProp(Node $node);
 
     /**
      * Visit a node with kind `\ast\AST_PROP_DECL`