+++ NEWS.md
@@ -17,6 +17,7 @@ New features (Analysis):
   `PhanMissingNamedArgument*`,
   `PhanDefinitelyDuplicateNamedArgument`, `PhanPositionalArgumentAfterNamedArgument`, and
   `PhanArgumentUnpackingUsedWithNamedArgument`
++ Incomplete support for analyzing uses of PHP 8.0's nullsafe operator(`?->`) for property reads and method calls. (#4067)
 
 Miscellaneous:
 + Raise the severity of `PhanUndeclaredConstant` from normal to critical.